Fix grouping issue
For some reasons items that are placed inside groups usually have the
group_id value as a numeric value instead of a string in those cases the
values instead of being in a grup are placed as individual items.

The following changes prevent the current scenario and add a fallback
mechanism for the previous mechanims in order to be backwards
compatible.